JW/gapago_langgraph/requirements.txt
# LangGraph
langgraph>=0.2.0

# Core
pydantic>=2.0.0
requests>=2.31.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0

# LLM Providers (선택적 설치)
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Exaone (로컬 모델용)
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
